requirements.txt
# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# HTTP client for external APIs
httpx>=0.25.2

# Ollama Python client
ollama>=0.1.6

# Environment variables
python-dotenv>=1.0.0

# Data validation
pydantic>=2.5.0

